AddedOtpVerificationToRegistration

+++ Controllers/v1.0/PassengerController.cs
@@ -31,7 +31,7 @@ namespace API.Controllers.v1
         public ActionResult updatePassenger(int id, PassengerUpdateDto passengerUpdateDto)
         {
             var passenger = _passengerRepository.GetPassengerById(id);
-            var user = _userRepository.GetUserById(passenger.UserId);
+            var user = _userRepository.GetUserById((int)passenger.UserId);
 
             if (passenger == null) return NotFound();
             _mapper.Map(passengerUpdateDto, passenger);
@@ -40,15 +40,6 @@ namespace API.Controllers.v1
 
             return BadRequest("Failed to Update Passenger");
         }
-
-        // [HttpGet("GetOTPs")]
-        // public ActionResult<List<OTP>> GetOTPs()
-        // {
-        //     var otps = _context.Passengers.Include(p => p.OTPs).FirstOrDefault(p => p.Id == 1).OTPs;
-        //     return otps;
-        // }
-
-
 
     }
 }
